requirements.txt
langchain-openai
langchain
python-dotenv
langgchain_community
langchain_core
streamlit